requirements.txt
PyQt6
sounddevice
numpy
soundfile
openai
scipy
pygame
google-generativeai
reportlab